Change the InventoryService.getItem signature
The item id type must always be long

+++ spark-inventory-service/src/main/java/org/example/spark/inventory/interactors/InventoryService.java
@@ -30,7 +30,7 @@ public interface InventoryService {
 
 	void updateAmount(long itemId, int amount);
 
-	RenderableItem getItem(int itemId);
+	RenderableItem getItem(long itemId);
 
 	RenderableItem[] getItems();
 }
